fix: adjust path for WidgetPicker
