+++ src/page/Editor/components/WidgetPickerEditor/components/ColorPicker/index.tsx
