+++ src/page/Editor/components/WidgetPickerEditor/components/ColorPicker/tests/color-picker.e2e.tsx
