+++ src/page/Editor/components/WidgetPickerEditor/components/InputRow/tests/label.e2e.tsx
