+++ src/page/Editor/components/WidgetPickerEditor/components/InputRow/tests/selectInput.e2e.tsx
