+++ src/page/Editor/components/WidgetPickerEditor/components/InputRow/tests/simpleInput.e2e.tsx
