+++ src/page/Editor/components/WidgetPickerEditor/components/InputRow/tests/switchInput.e2e.tsx
