+++ src/wrappedComponents/Wrapper/interface.tsx
@@ -1,0 +1,3 @@
+import { HTMLAttributes } from "react"
+
+export interface WrapperProps extends HTMLAttributes<HTMLDivElement> {}
